rename @repo/storage-chrome
