+++ apps/extension/package.json
@@ -35,7 +35,7 @@
     "@penumbra-zone/wasm": "50.0.0",
     "@radix-ui/react-icons": "^1.3.2",
     "@repo/context": "workspace:*",
-    "@repo/prax-storage": "workspace:*",
+    "@repo/storage-chrome": "workspace:*",
     "@repo/ui": "workspace:*",
     "@tanstack/react-query": "5.55.0",
     "exponential-backoff": "^3.1.1",
