+++ apps/extension/src/ctx/full-viewing-key.ts
@@ -1,6 +1,6 @@
 import { FullViewingKey } from '@penumbra-zone/protobuf/penumbra/core/keys/v1/keys_pb';
 import { Code, ConnectError } from '@connectrpc/connect';
-import { localExtStorage } from '@repo/prax-storage/local';
+import { localExtStorage } from '@repo/storage-chrome/local';
 
 export const getFullViewingKey = async () => {
   const wallet0 = (await localExtStorage.get('wallets'))[0];
