+++ apps/extension/src/routes/page/index.tsx
@@ -2,7 +2,7 @@ import { redirect } from 'react-router-dom';
 import { PagePath } from './paths';
 import { SplashPage } from '@repo/ui/components/ui/splash-page';
 import { Button } from '@repo/ui/components/ui/button';
-import { localExtStorage } from '@repo/prax-storage/local';
+import { localExtStorage } from '@repo/storage-chrome/local';
 import { useStore } from '../../state';
 import { getDefaultFrontend } from '../../state/default-frontend';
