+++ apps/extension/src/routes/popup/settings/settings-connected-sites/index.tsx
@@ -1,7 +1,7 @@
 import { MagnifyingGlassIcon } from '@radix-ui/react-icons';
 import { Input } from '@repo/ui/components/ui/input';
 import { LinkGradientIcon } from '../../../../icons/link-gradient';
-import { OriginRecord } from '@repo/prax-storage/types';
+import { OriginRecord } from '@repo/storage-chrome/types';
 import { AllSlices, useStore } from '../../../../state';
 import { UserChoice } from '@penumbra-zone/types/user-choice';
 import { SettingsScreen } from '../settings-screen';
