+++ apps/extension/src/routes/popup/settings/settings-connected-sites/known-site.tsx
@@ -1,4 +1,4 @@
-import { OriginRecord } from '@repo/prax-storage/types';
+import { OriginRecord } from '@repo/storage-chrome/types';
 import { UserChoice } from '@penumbra-zone/types/user-choice';
 import { Button } from '@repo/ui/components/ui/button';
 import { TrashIcon } from 'lucide-react';
