+++ apps/extension/src/rpc/index.ts
@@ -24,7 +24,7 @@ import { sctImpl } from '@penumbra-zone/services/sct-service';
 import { stakeImpl } from '@penumbra-zone/services/stake-service';
 import { viewImpl } from '@penumbra-zone/services/view-service';
 import { createProxyImpl, noContextHandler } from '@penumbra-zone/transport-dom/proxy';
-import { onboardGrpcEndpoint } from '@repo/prax-storage/onboard';
+import { onboardGrpcEndpoint } from '@repo/storage-chrome/onboard';
 import { rethrowImplErrors } from './rethrow-impl-errors';
 
 type RpcImplTuple<T extends ServiceType> = [T, Partial<ServiceImpl<T>>];
