+++ apps/extension/src/state/default-frontend.ts
@@ -1,6 +1,6 @@
 import { AllSlices, SliceCreator } from '.';
-import { ExtensionStorage } from '@repo/prax-storage/base';
-import { LocalStorageState } from '@repo/prax-storage/types';
+import { ExtensionStorage } from '@repo/storage-chrome/base';
+import { LocalStorageState } from '@repo/storage-chrome/types';
 
 export interface DefaultFrontendSlice {
   url?: string;
