+++ apps/extension/src/state/index.ts
@@ -5,10 +5,10 @@ import { customPersist } from './persist';
 import { createPasswordSlice, PasswordSlice } from './password';
 import { createSeedPhraseSlice, SeedPhraseSlice } from './seed-phrase';
 import { createNetworkSlice, NetworkSlice } from './network';
-import { localExtStorage } from '@repo/prax-storage/local';
-import { LocalStorageState } from '@repo/prax-storage/types';
-import { sessionExtStorage, SessionStorageState } from '@repo/prax-storage/session';
-import { ExtensionStorage } from '@repo/prax-storage/base';
+import { localExtStorage } from '@repo/storage-chrome/local';
+import { LocalStorageState } from '@repo/storage-chrome/types';
+import { sessionExtStorage, SessionStorageState } from '@repo/storage-chrome/session';
+import { ExtensionStorage } from '@repo/storage-chrome/base';
 import { createTxApprovalSlice, TxApprovalSlice } from './tx-approval';
 import { createOriginApprovalSlice, OriginApprovalSlice } from './origin-approval';
 import { ConnectedSitesSlice, createConnectedSitesSlice } from './connected-sites';
