+++ apps/extension/src/state/password.ts
@@ -1,8 +1,8 @@
 import { AllSlices, SliceCreator } from '.';
 import { Key, KeyJson, KeyPrint } from '@penumbra-zone/crypto-web/encryption';
-import { ExtensionStorage } from '@repo/prax-storage/base';
-import { LocalStorageState } from '@repo/prax-storage/types';
-import { SessionStorageState } from '@repo/prax-storage/session';
+import { ExtensionStorage } from '@repo/storage-chrome/base';
+import { LocalStorageState } from '@repo/storage-chrome/types';
+import { SessionStorageState } from '@repo/storage-chrome/session';
 
 // Documentation in /docs/custody.md
