+++ apps/extension/src/state/wallets.test.ts
@@ -2,9 +2,9 @@ import { create, StoreApi, UseBoundStore } from 'zustand';
 import { AllSlices, initializeStore } from '.';
 import { beforeEach, describe, expect, test, vi } from 'vitest';
 import { webcrypto } from 'crypto';
-import { ExtensionStorage } from '@repo/prax-storage/base';
-import { LocalStorageState } from '@repo/prax-storage/types';
-import { mockLocalExtStorage, mockSessionExtStorage } from '@repo/prax-storage/mock';
+import { ExtensionStorage } from '@repo/storage-chrome/base';
+import { LocalStorageState } from '@repo/storage-chrome/types';
+import { mockLocalExtStorage, mockSessionExtStorage } from '@repo/storage-chrome/mock';
 import type { WalletCreate } from '@penumbra-zone/types/wallet';
 
 vi.stubGlobal('crypto', webcrypto);
