+++ apps/extension/src/wallet-services.ts
@@ -3,8 +3,8 @@ import { AppService } from '@penumbra-zone/protobuf';
 import { createGrpcWebTransport } from '@connectrpc/connect-web';
 import { createClient } from '@connectrpc/connect';
 import { FullViewingKey, WalletId } from '@penumbra-zone/protobuf/penumbra/core/keys/v1/keys_pb';
-import { localExtStorage } from '@repo/prax-storage/local';
-import { onboardGrpcEndpoint, onboardWallet } from '@repo/prax-storage/onboard';
+import { localExtStorage } from '@repo/storage-chrome/local';
+import { onboardGrpcEndpoint, onboardWallet } from '@repo/storage-chrome/onboard';
 import { Services } from '@repo/context';
 import { WalletServices } from '@penumbra-zone/types/services';
 import { AssetId } from '@penumbra-zone/protobuf/penumbra/core/asset/v1/asset_pb';
