+++ packages/storage-chrome/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@repo/prax-storage",
+  "name": "@repo/storage-chrome",
   "version": "0.0.1",
   "private": true,
   "license": "(MIT OR Apache-2.0)",
@@ -12,11 +12,11 @@
     "./*": "./src/*.ts"
   },
   "devDependencies": {
-    "@penumbra-labs/registry": "^12.5.1",
-    "@penumbra-zone/bech32m": "15.0.0",
-    "@penumbra-zone/crypto-web": "41.0.1",
-    "@penumbra-zone/protobuf": "9.0.0",
-    "@penumbra-zone/types": "32.2.1",
+    "@penumbra-labs/registry": "^12.7.0",
+    "@penumbra-zone/bech32m": "17.0.0",
+    "@penumbra-zone/crypto-web": "44.0.0",
+    "@penumbra-zone/protobuf": "10.1.0",
+    "@penumbra-zone/types": "34.0.0",
     "@types/chrome": "0.0.270",
     "@types/lodash": "^4.17.13",
     "immer": "^10.1.1",
