+++ packages/storage-chrome/src/base-defaults.test.ts
