+++ packages/storage-chrome/src/base.ts
