+++ packages/storage-chrome/src/mock.ts
