+++ packages/storage-chrome/src/types.ts
